SupplementaryView-Demo-Standardization
This change will display (circle) the supplementaryView on three
specific days in the current month/year.  The 3rd, 13th and 23rd, are
expected to be circled when launched.  Other months/years, show random
days displaying the supplementaryView as in previous demos.

Since a bug was found in the Demo that was “circling” the wrong dates,
this will ensure that future changes always show these three dates
circled in the Demo